Use public-clouds to make config less boilerplate
This also updates the cloud names to be a bit terser. It moves them away
from being 1:1 with region names, but since we're also switching
projects etc I think it makes more sense to name them like this. Also
defaulting to sandbox is probably a good thing, only use the admin creds
when you really need to.