Simplify credential management
We don't need to store passwords anymore now that we have the replacekey
helper. Instead, just expose the Chameleon password, which we _do_ still
need because it's how you authenticate to Keystone.

Also add support for different ANL username and change the login gateway
for UC nodes to be the MCS login gateway.