Update resource provider status script
Nodes in maintenance mode will show up as reserved, but that's OK. But,
sometimes a node has inventory reserved, but no allocations or anything.
That is weird, and we should inform the operator.

Turns out, having nodes that are marked as used, but _not_ reserved-that
is OK. Seems to happen when an instance is running on a node already.