Add config merging script
The cc-openstack container now takes a dependency on pyyaml.

This patch now is able to take in multiple configs at run-time and merge them into 1. This way, users can simply set the environment variable and no additional installation steps are required.

The cloud-config-merge script is a python script which is copied into the cc-openstack image.

Configs are detected by the cc script, mounted as volumes to the cc-openstack container in the /etc/openstack/clouds.d directory, and then the entrypoint script runs cloud-config-merge with all of the configs as arguments.

cloud-config-merge writes the merged config to /etc/openstack/clouds[-public].yaml